# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CortexCloud API",
    "version": "1.0.0",
    "description": "Production-ready OpenAI-compatible AI gateway with x402 payment-gated routes. Agents pay per call in USDC on Base (eip155:8453) via the x402 protocol \u2014 no API key or subscription required.",
    "contact": {
      "email": "ampwerajonathan50@gmail.com"
    },
    "x-guidance": "CortexCloud is an agent-first AI inference + data gateway. Every paid endpoint is protected by the x402 payment protocol: send your request and the server replies 402 with a PaymentRequirements challenge (USDC on Base mainnet, eip155:8453). Pay the exact amount client-side with an x402-compatible wallet, then resend the same request with the X-PAYMENT header set to the base64 payment proof. On success you receive the normal JSON response. No API key, signup, or subscription is required \u2014 payment IS authentication. Paid routes: POST /x402/v1/chat/completions ($0.005), POST /x402/v1/responses ($0.005), POST /x402/v1/embeddings ($0.001), DATA marketplace GET /x402/v1/data/prices, /data/coins/search, /data/dex/search, /data/dex/pairs (each $0.001, CoinGecko/DEXScreener, no caller key), and ON-CHAIN Base data GET /x402/v1/data/base/balance, /data/base/token-balance, /data/base/nonce (each $0.001, public Base RPC, keyless). The free GET /x402/v1/models lists all 23+ models (Gemini, Groq, OpenRouter, NVIDIA NIM). Use model ids exactly as returned by /models. Set stream:true on chat for SSE. Discover and pay via x402/Bazaar; MCP endpoint at /x402/v1/mcp."
  },
  "servers": [
    {
      "url": "https://api.cortexcloud.org"
    }
  ],
  "paths": {
    "/x402/v1/chat/completions": {
      "post": {
        "operationId": "chat_completions",
        "summary": "OpenAI-compatible chat completion (x402 payment-gated)",
        "description": "Send an OpenAI-style chat completion request. Returns HTTP 402 with an x402 PaymentRequirements challenge (USDC on Base) unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "stream": {
                    "type": "boolean"
                  },
                  "temperature": {
                    "type": "number"
                  },
                  "max_tokens": {
                    "type": "integer"
                  }
                },
                "required": [
                  "model",
                  "messages"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/responses": {
      "post": {
        "operationId": "responses",
        "summary": "OpenAI Responses API completion (x402 payment-gated)",
        "description": "OpenAI Responses API endpoint. Returns HTTP 402 with an x402 PaymentRequirements challenge (USDC on Base) unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "input": {
                    "type": "string"
                  },
                  "stream": {
                    "type": "boolean"
                  },
                  "temperature": {
                    "type": "number"
                  },
                  "max_output_tokens": {
                    "type": "integer"
                  }
                },
                "required": [
                  "model",
                  "input"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/embeddings": {
      "post": {
        "operationId": "embeddings",
        "summary": "OpenAI-compatible text embeddings (x402 payment-gated)",
        "description": "Generate text embeddings. Returns HTTP 402 with an x402 PaymentRequirements challenge (USDC on Base) unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  }
                },
                "required": [
                  "input"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/prices": {
      "get": {
        "operationId": "data_prices",
        "summary": "Spot token prices (x402 payment-gated, CoinGecko)",
        "description": "Real-time spot prices from CoinGecko. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied. No caller-side API key required.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated CoinGecko coin ids, e.g. bitcoin,ethereum"
          },
          {
            "name": "vs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated vs currencies, e.g. usd,eur"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/coins/search": {
      "get": {
        "operationId": "data_coin_search",
        "summary": "Search coins (x402 payment-gated, CoinGecko)",
        "description": "Search CoinGecko coins by name/symbol. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Coin name or symbol to search"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/dex/search": {
      "get": {
        "operationId": "data_dex_search",
        "summary": "Search DEX pairs (x402 payment-gated, DEXScreener)",
        "description": "Search DEX pairs/liquidity on DEXScreener by token or address. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token symbol or address to search, e.g. WETH"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/dex/pairs": {
      "get": {
        "operationId": "data_dex_pairs",
        "summary": "Top DEX pairs for a token (x402 payment-gated, DEXScreener)",
        "description": "Top DEX pairs for a token on a given chain via DEXScreener. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "chain",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Chain id, e.g. ethereum, base, solana"
          },
          {
            "name": "pair",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token address or pair address"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/base/balance": {
      "get": {
        "operationId": "base_balance",
        "summary": "Native ETH balance on Base (x402 payment-gated, public RPC)",
        "description": "Native ETH balance of a Base address via public Base JSON-RPC. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied. Keyless.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Base wallet address, e.g. 0x1234...abcd"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/base/token-balance": {
      "get": {
        "operationId": "base_token_balance",
        "summary": "ERC-20 token balance on Base (x402 payment-gated, public RPC)",
        "description": "ERC-20 token balance for a Base address via public Base JSON-RPC (balanceOf). token may be an address or symbol (usdc, weth, dai, usdt). Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied. Keyless.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Base wallet address"
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token address OR symbol (usdc, weth, dai, usdt)"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/base/nonce": {
      "get": {
        "operationId": "base_nonce",
        "summary": "Account nonce on Base (x402 payment-gated, public RPC)",
        "description": "Transaction count (nonce) for a Base address via public Base JSON-RPC. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied. Keyless.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Base wallet address"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/models": {
      "get": {
        "operationId": "list_models",
        "summary": "List available models (free, no payment required)",
        "description": "Returns the available model registry. Free endpoint \u2014 no x402 payment required.",
        "security": [],
        "responses": {
          "200": {
            "description": "Model list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "owned_by": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}